free store, put types to defined types.d.ts file
